requirements.txt
uvicorn
fastapi
python-dotenv
pydantic-settings
httpx
pinecone
pinecone-text
motor
python-multipart
google-auth-oauthlib
google-api-python-client
google-genai
vulture
autoflake
isort
black